LIGHTSOFF can be decimal duration

+++ ScriptedEvents/Actions/LightsOffAction.cs
@@ -42,7 +42,7 @@
             if (!ScriptHelper.TryGetRooms(Arguments[0], out Room[] rooms, script))
                 return new(MessageType.NoRoomsFound, this, "rooms", Arguments[0]);
 
-            if (!VariableSystem.TryParse(Arguments[1], out int duration, script))
+            if (!VariableSystem.TryParse(Arguments[1], out float duration, script))
             {
                 return new(MessageType.NotANumber, this, "duration", Arguments[1]);
             }
